[Enhancement] Fix FE sonar bugs (#14053)

+++ fe/fe-core/src/main/java/com/starrocks/common/io/FastByteArrayInputStream.java
@@ -68,7 +68,7 @@ public class FastByteArrayInputStream extends InputStream {
 
     public final long skip(long n) {
         if ((pos + n) > count) {
-            n = count - pos;
+            n = (long) count - (long) pos;
         }
         if (n < 0) {
             return 0;
